// src/types/createCar.ts
export interface createCarType {
  title: string;
  carStatus: string;
  price: number | null;
  company: string;
  model: string;
  years: number | null;
  work: number | null;
  platform: string;
  fuel: string;
  gearbox: string;
  color: string;
  description: string;
  firstname: string;
  lastname: string;
  phone: number | null;
}